+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-insight-analytics-actions.ts
@@ -9,6 +9,7 @@ import {
   AnalyticsType,
   makeInsightAnalyticsAction,
 } from '../../analytics/analytics-utils';
+import {getCaseContextAnalyticsMetadata} from '../../case-context/case-context-state';
 import {facetIdDefinition} from '../generic/facet-actions-validation';
 import {LogCategoryFacetBreadcrumbActionCreatorPayload} from './category-facet-set-analytics-actions';
 
@@ -46,9 +47,11 @@ export const logCategoryFacetBreadcrumb = (
     AnalyticsType.Search,
     (client, state) => {
       validatePayload(payload, categoryFacetBreadcrumbPayloadDefinition);
+      const metadata = {
+        ...getCategoryFacetMetadata(state, payload),
+        ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
+      };
 
-      return client.logBreadcrumbFacet(
-        getCategoryFacetMetadata(state, payload)
-      );
+      return client.logBreadcrumbFacet(metadata);
     }
   )();