+++ packages/headless/src/features/facets/generic/facet-generic-insight-analytics-actions.ts
@@ -2,12 +2,15 @@ import {
   AnalyticsType,
   makeInsightAnalyticsAction,
 } from '../../analytics/analytics-utils';
+import {getCaseContextAnalyticsMetadata} from '../../case-context/case-context-state';
 
 export const logClearBreadcrumbs = () =>
   makeInsightAnalyticsAction(
     'analytics/facet/deselectAllBreadcrumbs',
     AnalyticsType.Search,
-    (client) => {
-      return client.logBreadcrumbResetAll();
+    (client, state) => {
+      return client.logBreadcrumbResetAll(
+        getCaseContextAnalyticsMetadata(state.insightCaseContext)
+      );
     }
   )();