+++ packages/headless/src/features/pagination/pagination-insight-analytics-actions.ts
@@ -3,6 +3,7 @@ import {
   AnalyticsType,
   makeInsightAnalyticsAction,
 } from '../analytics/analytics-utils';
+import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 import {currentPageSelector} from './pagination-selectors';
 
 export const logPageNumber = makeInsightAnalyticsAction(
@@ -11,6 +12,7 @@ export const logPageNumber = makeInsightAnalyticsAction(
   (client, state) =>
     client.logPagerNumber({
       pagerNumber: currentPageSelector(state as PaginationSection),
+      ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
     })
 );
 
@@ -20,6 +22,7 @@ export const logPageNext = makeInsightAnalyticsAction(
   (client, state) =>
     client.logPagerNext({
       pagerNumber: currentPageSelector(state as PaginationSection),
+      ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
     })
 );
 
@@ -29,5 +32,6 @@ export const logPagePrevious = makeInsightAnalyticsAction(
   (client, state) =>
     client.logPagerPrevious({
       pagerNumber: currentPageSelector(state as PaginationSection),
+      ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
     })
 );