+++ packages/samples/atomic-react/src/pages/ResultListPage.tsx
@@ -43,6 +43,17 @@ export const ResultListPage: FunctionComponent = () => {
 function MyTemplate(result: Result) {
   return (
     <>
+      <style>{`
+        .field {
+          display: inline-flex;
+          align-items: center;
+        }
+
+        .field-label {
+          font-weight: bold;
+          margin-right: 0.25rem;
+        }
+      `}</style>
       <AtomicResultSectionBadges>
         <AtomicResultBadge field="ec_brand" />
       </AtomicResultSectionBadges>
@@ -66,30 +77,33 @@ function MyTemplate(result: Result) {
       </AtomicResultSectionExcerpt>
       <AtomicResultSectionBottomMetadata>
         <AtomicResultFieldsList>
-          <AtomicResultDate format="ddd MMM D YYYY" />
+          <div className="field">
+            <AtomicText value="Date" />
+            <AtomicResultDate format="ddd MMM D YYYY" />
+          </div>
           {result.raw.cat_platform !== undefined && (
-            <>
+            <div className="field">
               <span className="field-label">
                 <AtomicText value="Platform" />
               </span>
               <AtomicResultText field="cat_platform" />
-            </>
+            </div>
           )}
           {result.raw.cat_condition !== undefined && (
-            <>
+            <div className="field">
               <span className="field-label">
                 <AtomicText value="Condition" />
               </span>
               <AtomicResultText field="cat_condition" />
-            </>
+            </div>
           )}
           {result.raw.cat_categories !== undefined && (
-            <>
+            <div className="field">
               <span className="field-label">
                 <AtomicText value="Tags" />
               </span>
               <AtomicResultMultiValueText field="cat_categories" />
-            </>
+            </div>
           )}
         </AtomicResultFieldsList>
       </AtomicResultSectionBottomMetadata>